+++ Runtime/SpriteSkinUtility.cs
@@ -79,6 +79,34 @@ namespace UnityEngine.U2D.Animation
             spriteSkin.boneTransforms = transforms;
         }
 
+        internal static int GetVertexStreamSize(this Sprite sprite)
+        {
+            int vertexStreamSize = 12;
+            if (sprite.HasVertexAttribute(Rendering.VertexAttribute.Normal))
+                vertexStreamSize = vertexStreamSize + 12;
+            if (sprite.HasVertexAttribute(Rendering.VertexAttribute.Tangent))
+                vertexStreamSize = vertexStreamSize + 16;
+            return vertexStreamSize;
+        }
+
+        internal static int GetVertexStreamOffset(this Sprite sprite, Rendering.VertexAttribute channel )
+        {
+            bool hasPosition = sprite.HasVertexAttribute(Rendering.VertexAttribute.Position);
+            bool hasNormals = sprite.HasVertexAttribute(Rendering.VertexAttribute.Normal);
+            bool hasTangents = sprite.HasVertexAttribute(Rendering.VertexAttribute.Tangent);
+
+            switch(channel)
+            {
+                case Rendering.VertexAttribute.Position:
+                    return hasPosition ? 0 : -1;
+                case Rendering.VertexAttribute.Normal:
+                    return hasNormals ? 12 : -1;
+                case Rendering.VertexAttribute.Tangent:
+                    return hasTangents ? (hasNormals ? 24 : 12) : -1;
+            }
+            return -1;
+        }
+
         private static void CreateGameObject(int index, SpriteBone[] spriteBones, Transform[] transforms, Transform root)
         {
             if (transforms[index] == null)
@@ -190,27 +218,71 @@ namespace UnityEngine.U2D.Animation
             return boneTransformHash;
         }
 
-        internal unsafe static void Deform(Matrix4x4 rootInv, NativeSlice<Vector3> vertices, NativeSlice<BoneWeight> boneWeights, NativeArray<Matrix4x4> boneTransforms, NativeSlice<Matrix4x4> bindPoses, NativeArray<Vector3> deformableVertices)
+        internal unsafe static void Deform(Sprite sprite, Matrix4x4 rootInv, NativeSlice<Vector3> vertices, NativeSlice<Vector4> tangents, NativeSlice<BoneWeight> boneWeights, NativeArray<Matrix4x4> boneTransforms, NativeSlice<Matrix4x4> bindPoses, NativeArray<byte> deformableVertices)
         {
             var verticesFloat3 = vertices.SliceWithStride<float3>();
+            var tangentsFloat4 = tangents.SliceWithStride<float4>();
             var bindPosesFloat4x4 = bindPoses.SliceWithStride<float4x4>();
+            var spriteVertexCount = sprite.GetVertexCount();
+            var spriteVertexStreamSize = sprite.GetVertexStreamSize();
             var boneTransformsFloat4x4 = NativeArrayUnsafeUtility.ConvertExistingDataToNativeArray<float4x4>(boneTransforms.GetUnsafePtr(), boneTransforms.Length, Allocator.None);
-            var deformableVerticesFloat3 = NativeArrayUnsafeUtility.ConvertExistingDataToNativeArray<float3>(deformableVertices.GetUnsafePtr(), deformableVertices.Length, Allocator.None);
+
+            byte* deformedPosOffset = (byte*)NativeArrayUnsafeUtility.GetUnsafePtr(deformableVertices);            
+            NativeSlice<float3> deformableVerticesFloat3 = NativeSliceUnsafeUtility.ConvertExistingDataToNativeSlice<float3>(deformedPosOffset, spriteVertexStreamSize, spriteVertexCount);
+            NativeSlice<float4> deformableTangentsFloat4 = NativeSliceUnsafeUtility.ConvertExistingDataToNativeSlice<float4>(deformedPosOffset, spriteVertexStreamSize, 1);   // Just Dummy.
+            if (sprite.HasVertexAttribute(Rendering.VertexAttribute.Tangent))
+            {
+                byte* deformedTanOffset = deformedPosOffset + sprite.GetVertexStreamOffset(Rendering.VertexAttribute.Tangent);
+                deformableTangentsFloat4 = NativeSliceUnsafeUtility.ConvertExistingDataToNativeSlice<float4>(deformedTanOffset, spriteVertexStreamSize, spriteVertexCount);
+            }
 
 #if ENABLE_UNITY_COLLECTIONS_CHECKS
             var handle1 = CreateSafetyChecks<float4x4>(ref boneTransformsFloat4x4);
             var handle2 = CreateSafetyChecks<float3>(ref deformableVerticesFloat3);
+            var handle3 = CreateSafetyChecks<float4>(ref deformableTangentsFloat4);
 #endif
 
-            Deform(rootInv, verticesFloat3, boneWeights, boneTransformsFloat4x4, bindPosesFloat4x4, deformableVerticesFloat3);
+            if (sprite.HasVertexAttribute(Rendering.VertexAttribute.Tangent))
+                Deform(rootInv, verticesFloat3, tangentsFloat4, boneWeights, boneTransformsFloat4x4, bindPosesFloat4x4, deformableVerticesFloat3, deformableTangentsFloat4);
+            else
+                Deform(rootInv, verticesFloat3, boneWeights, boneTransformsFloat4x4, bindPosesFloat4x4, deformableVerticesFloat3);
 
 #if ENABLE_UNITY_COLLECTIONS_CHECKS
             DisposeSafetyChecks(handle1);
             DisposeSafetyChecks(handle2);
+            DisposeSafetyChecks(handle3);
 #endif
         }
 
-        internal static void Deform(float4x4 rootInv, NativeSlice<float3> vertices, NativeSlice<BoneWeight> boneWeights, NativeArray<float4x4> boneTransforms, NativeSlice<float4x4> bindPoses, NativeArray<float3> deformed)
+        internal static void Deform(float4x4 rootInv, NativeSlice<float3> vertices, NativeSlice<BoneWeight> boneWeights, NativeArray<float4x4> boneTransforms, NativeSlice<float4x4> bindPoses, NativeSlice<float3> deformed)
+        {
+            if (boneTransforms.Length == 0)
+                return;
+
+            for (var i = 0; i < boneTransforms.Length; i++)
+            {
+                var bindPoseMat = bindPoses[i];
+                var boneTransformMat = boneTransforms[i];
+                boneTransforms[i] = math.mul(rootInv, math.mul(boneTransformMat, bindPoseMat));
+            }
+
+            for (var i = 0; i < vertices.Length; i++)
+            {
+                var bone0 = boneWeights[i].boneIndex0;
+                var bone1 = boneWeights[i].boneIndex1;
+                var bone2 = boneWeights[i].boneIndex2;
+                var bone3 = boneWeights[i].boneIndex3;
+
+                var vertex = vertices[i];
+                deformed[i] =
+                    math.transform(boneTransforms[bone0], vertex) * boneWeights[i].weight0 +
+                    math.transform(boneTransforms[bone1], vertex) * boneWeights[i].weight1 +
+                    math.transform(boneTransforms[bone2], vertex) * boneWeights[i].weight2 +
+                    math.transform(boneTransforms[bone3], vertex) * boneWeights[i].weight3;
+            }
+        }
+
+        internal static void Deform(float4x4 rootInv, NativeSlice<float3> vertices, NativeSlice<float4> tangents, NativeSlice<BoneWeight> boneWeights, NativeArray<float4x4> boneTransforms, NativeSlice<float4x4> bindPoses, NativeSlice<float3> deformed, NativeSlice<float4> deformedTangents)
         {
             if(boneTransforms.Length == 0)
                 return;
@@ -228,21 +300,33 @@ namespace UnityEngine.U2D.Animation
                 var bone1 = boneWeights[i].boneIndex1;
                 var bone2 = boneWeights[i].boneIndex2;
                 var bone3 = boneWeights[i].boneIndex3;
-
+                
+                var vertex = vertices[i];
                 deformed[i] =
-                    math.transform(boneTransforms[bone0], vertices[i]) * boneWeights[i].weight0 +
-                    math.transform(boneTransforms[bone1], vertices[i]) * boneWeights[i].weight1 +
-                    math.transform(boneTransforms[bone2], vertices[i]) * boneWeights[i].weight2 +
-                    math.transform(boneTransforms[bone3], vertices[i]) * boneWeights[i].weight3;
+                    math.transform(boneTransforms[bone0], vertex) * boneWeights[i].weight0 +
+                    math.transform(boneTransforms[bone1], vertex) * boneWeights[i].weight1 +
+                    math.transform(boneTransforms[bone2], vertex) * boneWeights[i].weight2 +
+                    math.transform(boneTransforms[bone3], vertex) * boneWeights[i].weight3;
+
+                var tangent = new float4(tangents[i].xyz, 0.0f);
+
+                tangent =
+                    math.mul(boneTransforms[bone0], tangent) * boneWeights[i].weight0 +
+                    math.mul(boneTransforms[bone1], tangent) * boneWeights[i].weight1 +
+                    math.mul(boneTransforms[bone2], tangent) * boneWeights[i].weight2 +
+                    math.mul(boneTransforms[bone3], tangent) * boneWeights[i].weight3;
+
+                deformedTangents[i] = new float4(math.normalize(tangent.xyz), tangents[i].w);
             }
         }
 
-        internal unsafe static void Deform(Sprite sprite, Matrix4x4 invRoot, Transform[] boneTransformsArray, ref NativeArray<Vector3> deformableVertices)
+        internal unsafe static void Deform(Sprite sprite, Matrix4x4 invRoot, Transform[] boneTransformsArray, ref NativeArray<byte> deformVertexData)
         {
             Debug.Assert(sprite != null);
-            Debug.Assert(sprite.GetVertexCount() == deformableVertices.Length);
+            Debug.Assert(sprite.GetVertexCount() == (deformVertexData.Length / sprite.GetVertexStreamSize()));
             
             var vertices = sprite.GetVertexAttribute<Vector3>(UnityEngine.Rendering.VertexAttribute.Position);
+            var tangents = sprite.GetVertexAttribute<Vector4>(UnityEngine.Rendering.VertexAttribute.Tangent);
             var boneWeights = sprite.GetVertexAttribute<BoneWeight>(UnityEngine.Rendering.VertexAttribute.BlendWeight);
             var bindPoses = sprite.GetBindPoses();
             
@@ -254,7 +338,7 @@ namespace UnityEngine.U2D.Animation
             for (var i = 0; i < boneTransformsArray.Length; ++i)
                 boneTransforms[i] = boneTransformsArray[i].localToWorldMatrix;
 
-            Deform(invRoot, vertices, boneWeights, boneTransforms, bindPoses, deformableVertices);
+            Deform(sprite, invRoot, vertices, tangents, boneWeights, boneTransforms, bindPoses, deformVertexData);
 
             boneTransforms.Dispose();
         }
@@ -269,43 +353,56 @@ namespace UnityEngine.U2D.Animation
             return handle;
         }
 
+        private static AtomicSafetyHandle CreateSafetyChecks<T>(ref NativeSlice<T> array) where T : struct
+        {
+            var handle = AtomicSafetyHandle.Create();
+            AtomicSafetyHandle.SetAllowSecondaryVersionWriting(handle, true);
+            AtomicSafetyHandle.UseSecondaryVersion(ref handle);
+            NativeSliceUnsafeUtility.SetAtomicSafetyHandle<T>(ref array, handle);
+            return handle;
+        }
+
         private static void DisposeSafetyChecks(AtomicSafetyHandle handle)
         {
             AtomicSafetyHandle.Release(handle);
         }
 #endif
 
-        internal static Vector3[] Bake(this SpriteSkin spriteSkin)
+        internal static void Bake(this SpriteSkin spriteSkin, ref NativeArray<byte> deformVertexData)
         {
             if (!spriteSkin.isValid)
                 throw new Exception("Bake error: invalid SpriteSkin");
 
             var sprite = spriteSkin.spriteRenderer.sprite;
             var boneTransformsArray = spriteSkin.boneTransforms;
-            var deformableVertices = new NativeArray<Vector3>(sprite.GetVertexCount(), Allocator.Temp, NativeArrayOptions.UninitializedMemory);
-
-            Deform(sprite, Matrix4x4.identity, boneTransformsArray, ref deformableVertices);
-            
-            var result = deformableVertices.ToArray();
-            deformableVertices.Dispose();
-            return result;
+            Deform(sprite, Matrix4x4.identity, boneTransformsArray, ref deformVertexData);
+           
         }
 
-        internal static void CalculateBounds(this SpriteSkin spriteSkin)
+        internal unsafe static void CalculateBounds(this SpriteSkin spriteSkin)
         {
             Debug.Assert(spriteSkin.isValid);
+            var sprite = spriteSkin.sprite;
+
+
+            var deformVertexData = new NativeArray<byte>(sprite.GetVertexStreamSize() * sprite.GetVertexCount(), Allocator.Temp, NativeArrayOptions.UninitializedMemory);
+            void* dataPtr = NativeArrayUnsafeUtility.GetUnsafePtr(deformVertexData);
+            var deformedPosSlice = NativeSliceUnsafeUtility.ConvertExistingDataToNativeSlice<Vector3>(dataPtr, sprite.GetVertexStreamSize(), sprite.GetVertexCount());
+#if ENABLE_UNITY_COLLECTIONS_CHECKS
+            NativeSliceUnsafeUtility.SetAtomicSafetyHandle(ref deformedPosSlice, NativeArrayUnsafeUtility.GetAtomicSafetyHandle(deformVertexData));
+#endif
 
             var rootBone = spriteSkin.rootBone;
-            var deformableVertices = spriteSkin.Bake();
+            spriteSkin.Bake(ref deformVertexData);
             var bounds = new Bounds();
 
-            if (deformableVertices.Length > 0)
+            if (deformVertexData.Length > 0)
             {
-                bounds.min = rootBone.InverseTransformPoint(deformableVertices[0]);
+                bounds.min = rootBone.InverseTransformPoint(deformedPosSlice[0]);
                 bounds.max = bounds.min;
             }
 
-            foreach(var v in deformableVertices)
+            foreach(var v in deformedPosSlice)
                 bounds.Encapsulate(rootBone.InverseTransformPoint(v));
 
             bounds.extents = Vector3.Scale(bounds.extents, new Vector3(1.25f, 1.25f, 1f));